requirements.txt
alembic==1.4.2
appnope==0.1.0
awscli==1.16.272
backcall==0.1.0
certifi==2020.4.5.1
click==7.1.1
decorator==4.4.2
docopt==0.6.2
entrypoints==0.3
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
get==2019.4.13
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.2
joblib==0.14.1
jupyter-client==6.1.2
jupyter-core==4.6.3
Mako==1.1.2
MarkupSafe==1.1.1
numpy==1.18.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
post==2019.4.13
prompt-toolkit==3.0.4
ptyprocess==0.6.0
public==2019.4.13
Pygments==2.6.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pythonanywhere==0.8.1
pyzmq==18.1.1
query-string==2019.4.13
request==2019.4.13
rsa==3.4.2
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
SQLAlchemy==1.3.16
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.9
Werkzeug==1.0.1
WTForms==2.2.1
gunicorn

scrapy
scrapy-user-agents
scrapy-proxies-tool